requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0
lxml>=4.9.0
webdriver-manager>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0
scikit-learn>=1.3.0
openpyxl>=3.1.0

